bugs

+++ lib/models/section.dart
@@ -75,7 +75,8 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems) {
-      if (!items.contains(original)) {
+      if (!items.contains(original) &&
+          (original.image as String).contains('firebase')) {
         try {
           final ref = storage.refFromURL(original.image as String);
           await ref.delete();
@@ -94,12 +95,15 @@ class Section extends ChangeNotifier {
 
   Future<void> delete() async {
     await firestoreRef.delete();
+
     for (final item in items) {
-      try {
-        final ref = storage.refFromURL(item.image as String);
-        await ref.delete();
-      } catch (e) {
-        debugPrint('Falha ao deletar $items');
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = storage.refFromURL(item.image as String);
+          await ref.delete();
+        } catch (e) {
+          debugPrint('Falha ao deletar $items');
+        }
       }
     }
   }
